Created an extra class and modify fragment
The class created is a custom listadapter that we use to set the images
for the message type and the user who sends it. In the inbox fragment we
initialized this and retrieved the data and now we are creating an
onclick listener for when the item is selected we can view it depending
on the file type.